first mission +fixes

+++ Project/Source/engine/Components/Cpps/Drawing.cpp
@@ -61,7 +61,7 @@ void Window::RecalculateSize()
 	glBindFramebuffer(GL_FRAMEBUFFER, NormalMapFBO);
 	glGenTextures(1, &NormalMapColorBuffer);
 	glBindTexture(GL_TEXTURE_2D, NormalMapColorBuffer);
-	glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA16F, ViewportSize.x, ViewportSize.y, 0, GL_RGBA, GL_FLOAT, NULL);
+	glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA32F, ViewportSize.x, ViewportSize.y, 0, GL_RGBA, GL_FLOAT, NULL);
 	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
 	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
 	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_S, GL_CLAMP_TO_EDGE);
